+++ class/chaumeilrfa.class.php
@@ -372,13 +372,7 @@ class ChaumeilRfa extends CommonObject
 		];
 		$classfortooltip = 'classfortooltip';
 		$dataparams = '';
-		if (getDolGlobalInt('MAIN_ENABLE_AJAX_TOOLTIP')) {
-			$classfortooltip = 'classforajaxtooltip';
-			$dataparams = ' data-params="'.dol_escape_htmltag(json_encode($params)).'"';
-			$label = '';
-		} else {
-			$label = implode($this->getTooltipContentArray($params));
-		}
+		$label = implode($this->getTooltipContentArray($params));
 		global $user;
 		if ($user->hasRight('clichaumeil', 'chaumeilrfa', 'write')) {
 			$url = dol_buildpath('/clichaumeil/chaumeilrfa_card.php', 1).'?id='.$this->id.'&action=edit&socid='.$this->fk_soc;
@@ -555,7 +549,7 @@ class ChaumeilRfa extends CommonObject
 
 		if (empty($this->labelStatus) || empty($this->labelStatusShort)) {
 			global $langs;
-			//$langs->load("clichaumeil@clichaumeil");
+			$langs->load("clichaumeil@clichaumeil");
 			$this->labelStatus[self::STATUS_DRAFT] = $langs->transnoentitiesnoconv('Draft');
 			$this->labelStatus[self::STATUS_WON] = $langs->transnoentitiesnoconv('RfaStatusWon');
 			$this->labelStatus[self::STATUS_LOST] = $langs->transnoentitiesnoconv('RfaStatusLost');